refactor(auth): new signup structure
