+++ frontend/src/app/auth/components/index.ts
@@ -1,0 +1,3 @@
+export { AuthForm } from './AuthForm'
+export { AuthHero } from './AuthHero'
+export { AuthSuggestion } from './AuthSuggestion'
