+++ frontend/src/app/auth/layout.tsx
@@ -1,7 +1,6 @@
 import type { PropsWithChildren } from 'react'
 
-import { AuthHero } from './components/AuthHero'
-import { AuthSuggestion } from './components/AuthSuggestion'
+import { AuthHero, AuthSuggestion } from './components'
 import styles from './layout.module.scss'
 
 export default function AuthLayout({ children }: PropsWithChildren) {
