+++ frontend/src/app/auth/signup/data/index.ts
@@ -1,0 +1,1 @@
+export { signupInputs } from './inputs'
