+++ frontend/src/app/auth/signup/interfaces/index.ts
@@ -1,0 +1,1 @@
+export type { SignupPayload } from './SignupPayload'
